+++ src/components/spartan/SInputOTP/index.ts
@@ -1,3 +1,4 @@
 export { default as SInputOTP } from './SInputOTP.vue';
+export { default as SInputOTPItem } from './SInputOTPItem.vue';
 
-export type { TInputOTPProps, TInputOTPEmits } from './types';
+export type { TInputOTPProps, TInputOTPEmits, TInputOTPItemProps } from './types';