+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- New Accounts OTP example.
 - New `SInputOTP` component.
+- New `SInputOTPItem` component.
 
 ## 2.0.0-beta.76 - 2024-03-04
 ### Fixed